+++ directory/src/FilterLists.Directory.Infrastructure/Persistence/Queries/Entities/Software.cs
@@ -6,7 +6,7 @@ namespace FilterLists.Directory.Infrastructure.Persistence.Queries.Entities
 {
     public class Software : IHaveSurrogateKey
     {
-        public ushort Id { get; private set; }
+        public int Id { get; private set; }
         public string Name { get; private set; } = null!;
         public string? Description { get; private set; }
         public Uri? HomeUrl { get; private set; }
