+++ directory/src/FilterLists.Directory.Infrastructure/Persistence/Queries/Entities/ViewUrlMirror.cs
@@ -5,7 +5,7 @@ namespace FilterLists.Directory.Infrastructure.Persistence.Queries.Entities
 {
     public class ViewUrlMirror : IHaveSurrogateKey
     {
-        public ushort Id { get; private set; }
+        public int Id { get; private set; }
         public FilterList FilterList { get; private set; } = null!;
         public Uri Url { get; private set; } = null!;
     }
